refactor: extend connectId → personalId rename to unit tests
Move all test files from org.commcare.connect.network.connectId to
org.commcare.connect.network.personalId and update package declarations
and imports. Completes the incomplete rename in unit test sources.
