+++ app/unit-tests/src/org/commcare/connect/network/personalId/parser/ConfirmBackupCodeResponseParserTest.kt
@@ -1,4 +1,4 @@
-package org.commcare.connect.network.connectId.parser
+package org.commcare.connect.network.personalId.parser
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.commcare.CommCareTestApplication
